feat: add hall of fame page
