+++ packages/ui/components/landing/Header/config.ts
@@ -2,7 +2,7 @@ export const MENU_ENTRIES = [
   { key: "/web/curriculum", text: "Currículo" },
   { key: "/web/recruitment", text: "Recrutamento" },
   //   { path: "/projects", text: "Projetos" },
-  { path: "/web/team", text: "Equipa" },
-  { path: "/blog", text: "Blog" },
-  { path: "/web/hall-of-fame", text: "Hall of Fame"}
+  { key: "/web/team", text: "Equipa" },
+  { key: "/web/hall-of-fame", text: "Hall of Fame" },
+  { key: "/blog", text: "Blog" },
 ];
